+++ backend/src/auth/auth.controller.ts
@@ -1,9 +1,9 @@
 import { Controller, Get, Res, UseGuards } from '@nestjs/common';
 import { NaverAuthGuard } from './guards/naverAuth.guard';
 import { User } from './utils/user.decorator';
-import { CreateUserDto } from './dto/createUserDto';
 import { AuthService } from './auth.service';
 import { Response } from 'express';
+import { CreateUserDto } from './dto/user.dto';
 
 @Controller('auth')
 export class AuthController {
