--- backend/src/auth/dto/createUserResponseDto.ts
@@ -1,4 +1,0 @@
-export class CreateUserResponseDto {
-  userId: number;
-  token: string;
-}
